% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-tshark.R
\name{find_tshark}
\alias{find_tshark}
\title{Find the \code{tshark} binary}
\usage{
find_tshark(path = Sys.getenv("TSHARK_PATH", ""))
}
\arguments{
\item{path}{hint to where to look for the \code{tshark} binary}
}
\value{
length 1 character vector of the path to the \code{tshark} binary or \code{""}
}
\description{
Use the environment variable \code{TSHARK_PATH} or specify the directory in
the call to this function.
}
\examples{
loc <- tryCatch(
  find_tshark(),
  error = function(e) message("No tshark")
)
}
