% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packet-sum.R
\name{packet_summary}
\alias{packet_summary}
\title{Extract packet summary table (if any) from a PCAP}
\usage{
packet_summary(pcap)
}
\arguments{
\item{pcap}{path to PCAP file (\code{\link[=path.expand]{path.expand()}} will be called on this value)}
}
\value{
data frame
}
\description{
Extract packet summary table (if any) from a PCAP
}
\examples{
tryCatch(
  packet_summary(system.file("pcap", "http.pcap", package = "tsharrk")),
  error = function(e) message("No tshark")
)
}
