% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcap-to-json.R
\name{pcap_to_json}
\alias{pcap_to_json}
\title{Convert an entire PCAP file to JSON}
\usage{
pcap_to_json(pcap, json, protos = c(), include_child_nodes = FALSE)
}
\arguments{
\item{pcap}{path to PCAP file (\code{\link[=path.expand]{path.expand()}} will be called on this value)}

\item{json}{path (including filename) to the location where you want the JSON
file stored (\code{\link[=path.expand]{path.expand()}} will be called on this value)}

\item{protos}{character vector of protocols to include. (default is all)}

\item{include_child_nodes}{if \code{protos} is specified, this logical parameter
(default \code{FALSE}) controls whether child nodes are included.}
}
\value{
(expanded) path to \code{json} (invisibly)
}
\description{
Convert an entire PCAP file to JSON
}
\examples{
tryCatch(
  pcap_to_json(system.file("pcap", "http.pcap", package = "tsharrk"), tempfile()),
  error = function(e) message("No tshark")
)
}
