% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-json.R
\name{ts_read_json}
\alias{ts_read_json}
\alias{print.tshark_json}
\title{Read in a JSON file created with \code{\link[=pcap_to_json]{pcap_to_json()}}}
\usage{
ts_read_json(json)

\method{print}{tshark_json}(x, ...)
}
\arguments{
\item{json}{path to JSON output from \code{\link[=pcap_to_json]{pcap_to_json()}}. \code{\link[=path.expand]{path.expand()}} will be
called on this value.}
}
\value{
\code{list} classed as \code{tshark_json} for better \code{print}ing.
}
\description{
Read in a JSON file created with \code{\link[=pcap_to_json]{pcap_to_json()}}
}
