% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tshark.R
\name{tshark_exec}
\alias{tshark_exec}
\title{Call the tshark binary with optional custom environment variables and options}
\usage{
tshark_exec(tshark_bin = find_tshark(), args = c(), env = c())
}
\arguments{
\item{tshark_bin}{specify a complete path or let \code{\link[=find_tshark]{find_tshark()}} do the dirty work.}

\item{args}{same as \code{\link[=system2]{system2()}} \code{args}}

\item{env}{same as \code{\link[=system2]{system2()}} \code{env}}
}
\value{
\code{list} with \code{stderr}, \code{stdout}, and \code{status} (invisibly)
}
\description{
This is just a convenience wrapper around \code{\link[=system2]{system2()}}. See \code{\link[=find_tshark]{find_tshark()}} for
information on helping this package find the tshark binary.
}
