% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsharrk-package.R
\docType{package}
\name{tsharrk}
\alias{tsharrk}
\alias{tsharrk-package}
\title{Tools to Make Analyses Using 'tshark' Easier}
\description{
The 'tshark' (\url{https://www.wireshark.org/docs/man-pages/tshark.html})
command line utility comes with Wireshark and is a is useful when performing
analyses on packet captures (PCAPs). Tools are provided to make it a bit easier
to work with 'tshark' to perform analyses with R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/tsharrk}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/tsharrk/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
